backend/requirements.txt
Flask==3.0.3
sqlalchemy==2.0.28
pyjwt==2.8.0
bcrypt==4.1.2
marshmallow-sqlalchemy==1.0.0
marshmallow==3.21.1
mysqlclient==2.2.4
flask_cors==4.0.0
python-pptx==0.6.23
python-dotenv==1.0.1
openai==1.14.0
pydub==0.25.1
waitress
mutagen